+++ src/App.jsx
@@ -1,83 +1,20 @@
-import { useEffect, useState } from "react"
 import Header from "./components/Header"
 import Guitar from "./components/Guitar"
-import { db} from "./data/db";
+import { useCart } from "./hooks/useCart";
 
 
 function App() {
-  const initialCart = () => {
-    const localStorageCart = localStorage.getItem('cart')
-    return localStorageCart ? JSON.parse(localStorageCart) : []
-  }
-  const[data] = useState(db);
-  const[cart, setCart] = useState(initialCart);
-
-  const MIN_ITEMS = 1
-  const MAX_ITEMS = 5
-
-  useEffect(() => {
-    localStorage.setItem('cart', JSON.stringify(cart))
-  }, [cart])
-
-  function addToCart(item) {
-    const itemExists = cart.findIndex(guitar => guitar.id === item.id)
-    if(itemExists >= 0 ) { 
-      if(cart[itemExists].quantity >= MAX_ITEMS) return
-      const updatedCart = [...cart]
-      updatedCart[itemExists].quantity++
-      setCart(updatedCart)
-    } else {
-      item.quantity = 1
-      setCart([...cart, item])
-    }
-  }
-
-  function removeFromCart(id){
-    setCart(prevCart => prevCart.filter(guitar => guitar.id !== id))
-  }
-
-  function increaseQuantity(id){
-    const updatedCart = cart.map(item =>{
-      if(item.id === id && item.quantity < MAX_ITEMS){
-        return{
-          ...item,
-          quantity: item.quantity + 1
-        }
-      }
-      return item
-    })
-    setCart(updatedCart)
-  }
-  function decreaseQuantity(id) {
-    const updatedCart = cart.map( item => {
-      if(item.id === id && item.quantity > MIN_ITEMS) {
-        return {
-          ...item,
-          quantity: item.quantity - 1
-        }
-      }
-      return item
-    })
-    setCart(updatedCart)
-  }
-
-  function clearCart(){
-    setCart([])
-  }
-
-  function saveLocalStorage(){
-    localStorage.setItem('cart', JSON.stringify(cart))
-  }
-
+  const {data,cart,addToCart,removeFromCart,decreaseQuantity,increaseQuantity,clearCart, isEmpty, cartTotal} = useCart()
   return (
-    <>
-    
+    <>    
     <Header 
       cart={cart}
       removeFromCart={removeFromCart}
       increaseQuantity={increaseQuantity}
       decreaseQuantity={decreaseQuantity}
       clearCart={clearCart}
+      isEmpty={isEmpty}
+      cartTotal={cartTotal}
     />
     
     <main class="container-xl mt-5">
@@ -88,7 +25,6 @@ function App() {
           <Guitar
             key={guitar.id}
             guitar={guitar}
-            setCart={setCart}
             addToCart={addToCart}
           />
           ))}
