+++ src/components/Header.jsx
@@ -1,11 +1,4 @@
-import { useMemo } from "react";
-import Guitar from "./Guitar";
-
-export default function Header({cart, removeFromCart, increaseQuantity, decreaseQuantity, clearCart}){
-
-    const isEmpty = useMemo( () => cart.length === 0, [cart])
-    const cartTotal = useMemo(() => cart.reduce( (total, item) => total + (item.quantity * item.price), 0), [cart])
-
+export default function Header({cart, removeFromCart, increaseQuantity, decreaseQuantity, clearCart, isEmpty, cartTotal}){    
     return(
         <header class="py-5 header">
         <div class="container-xl">
